Revert "Feature/user model"
