--- app/api/api_v1/views/__init__.py
@@ -1,3 +1,0 @@
-"""
-Main views package.
-"""
