--- app/crud/__init__.py
@@ -1,1 +1,0 @@
-from .crud_user import user
