--- app/db/__init__.py
@@ -1,4 +1,0 @@
-"""
-Database package.
-"""
-from app.db.base import Base
