--- app/models/__init__.py
@@ -1,7 +1,0 @@
-"""
-SQLAlchemy models.
-"""
-from app.db import Base
-from app.models.user import User
-
-__all__ = ["User", "Base"]
