--- app/schemas/__init__.py
@@ -1,4 +1,0 @@
-"""
-Pydantic schemas package.
-"""
-from .user import User, UserCreate, UserUpdate, UserInDB
