+++ bshr/api/api_v1/api.py
@@ -1,0 +1,6 @@
+from fastapi import APIRouter
+
+from bshr.bshr.api.api_v1.views import login
+
+api_router = APIRouter()
+api_router.include_router(login.router, tags=["login"])
