+++ bshr/crud/base.py
@@ -1,6 +1,3 @@
-"""
-Common CRUD methods.
-"""
 from typing import TypeVar, Type, Generic, Optional, Any, List, Union, Dict
 
 from fastapi.encoders import jsonable_encoder
@@ -8,7 +5,7 @@ from pydantic import BaseModel
 from sqlalchemy.future import select
 from sqlalchemy.orm import Session
 
-from app.db.base_class import Base
+from db.base_class import Base
 
 ModelType = TypeVar("ModelType", bound=Base)
 CreateSchemaType = TypeVar("CreateSchemaType", bound=BaseModel)
@@ -32,7 +29,7 @@ class CRUDBase(Generic[ModelType, CreateSchemaType, UpdateSchemaType]):
         :param id: object id
         :return: SQLAlchemy model class
         """
-        res = await db.execute(select(self.model).filter(self.model.id == id))
+        res = await db.execute(select(self.model)).filter(self.model.id == id)
         found_obj = res.scalar_one_or_none()
         return found_obj
 
@@ -101,7 +98,7 @@ class CRUDBase(Generic[ModelType, CreateSchemaType, UpdateSchemaType]):
         :param id: object id
         :return: SQLAlchemy model class
         """
-        res = await db.execute(select(self.model).filter(self.model.id == id))
+        res = await db.execute(select(self.model)).filter(self.model.id == id)
         found_obj = res.scalar_one_or_none()
         await db.delete(found_obj)
         await db.commit()
