+++ bshr/db/base.py
@@ -2,4 +2,4 @@
 Import all the models, so that Base has them before being
 imported by Alembic
 """
-from app.db.base_class import Base  # noqa
+from db.base_class import Base  # noqa
