+++ bshr/db/base_class.py
