+++ pyproject.toml
@@ -9,15 +9,14 @@ readme = "README.md"
 [tool.poetry.dependencies]
 python = ">=3.8.1,<4.0"
 fastapi = "^0.88.0"
-sqlalchemy = {extras = ["asyncio"], version = "^1.4.45"}
+sqlalchemy = "^1.4.45"
 greenlet = "^2.0.1"
 hypercorn = {extras = ["uvloop"], version = "^0.14.3"}
 alembic = "^1.8.1"
-pydantic = {extras = ["email"], version = "^1.10.2"}
+pydantic = "^1.10.2"
 passlib = "^1.7.4"
 python-jose = "^3.3.0"
 uvloop = "^0.17.0"
-asyncpg = "^0.27.0"
 
 
 [tool.poetry.group.dev.dependencies]
